net/ntp: Fix stage when devel/bitkeeper is installed
NTP is developed using the bitkeeper VCS. checkHtmlFileDates, uses
bitkeeper metadata to alter dates in html files. This results in
a bunch of *.old files installed in ${STAGEDIR}. As the distribution
tarball contains no bitkeeper metadata, no dates are updated at the
cost of failed build. This patch teaches scripts/build/checkHtmlFileDates
to act as if bitkeeper is not installed regardless of bitkeeper's
installatikon status.

MFH:		2021Q4