devel/py-pythran: Add new port
Pythran is an ahead of time compiler for a subset of the Python language,
with a focus on scientific computing. It takes a Python module annotated
with a few interface descriptions and turns it into a native Python module
with the same interface, but (hopefully) faster.

WWW: https://github.com/serge-sans-paille/pythran