devel/py-breathe: RST&Sphinx bridge to Doxygen [NEW PORT]
Breathe provides a bridge between the Sphinx and Doxygen documentation systems.

It is an easy way to include Doxygen information in a set of documentation
generated by Sphinx. The aim is to produce an autodoc like support for people
who enjoy using Sphinx but work with languages other than Python. The system
relies on the Doxygen’s xml output.

WWW: https://github.com/michaeljones/breathe