lang/gcc8: Relinquish maintainership
After working on (and largely maintaining) our GCC ports for more than
19 years, time has come to hand over the baton. Sadly despite multiple
requests nobody stepped up, so return this port to the pool.

That said, the GCC 8 release series is end of life for half a year now
and this port should be removed rather sooner than later in favor of
lang/gcc10 or later. We only keep it for the sake of a sole other port,
converters/wkhtmltopdf.