devel/sdbus-cpp: add new port
sdbus-c++ is a high-level C++ D-Bus library for Linux designed to
provide expressive, easy-to-use API in modern C++. It adds another
layer of abstraction on top of sd-bus, a nice, fresh C D-Bus
implementation by systemd.

sdbus-c++ has been written primarily as a replacement of dbus-c++,
which currently suffers from a number of (unresolved) bugs,
concurrency issues and inherent design complexities and
limitations. sdbus-c++ has learned from dbus-c++ and has chosen a
different path, a path of simple yet powerful design that is intuitive
and friendly to the user and inherently free of those bugs.

Even though sdbus-c++ uses sd-bus library, it is not necessarily
constrained to systemd and can perfectly be used in non-systemd
environments as well.

https://github.com/Kistler-Group/sdbus-cpp