www/py-adblock: fix build on powerpc64*
Python uses architecture name from "uname -m", but on FreeBSD it should
actually use "uname -p".

This causes pip to misrecognize valid wheel as wrong:
ERROR: adblock-0.5.0-cp36-abi3-freebsd_13_0_RELEASE_p4_powerpc64.whl is not a supported wheel on this platform.

This patch renames files on powerpc64* to end with powerpc.whl, which is acceptable for pip.