math/fastops: new port had been added (+)
This small library enables acceleration of bulk calls of certain math
functions on AVX and AVX2 hardware.  Currently supported operations are
exp, log, sigmoid, and tanh.  The library is designed with extensibility
in mind.

Restrict the port to amd64 as it's inherently useless anywhere else and
does not even build at least on i386.

WWW: https://github.com/yandex/fastops