devel/nimble: Reinstate port and update to 0.13.1
The official documentation of nim states that nimble is included with
the distribution of nim, but this is somewhat disingenuous as koch,
the nim builder, will fetch the nimble source code from git. So
reinstate devel/nimble as separate port of nimble, which is not built
by lang/nim.

PR:		259522