www/gitlab-ce: security upgrade to 14.4.1
This version has a problem showing files like .md files (e.g. README.md on the start page of a project).
But as this bug is already present on version 14.4.0
I will deploy this now, to fix the security problems.
We need to find the source why this happens and fix it later.

Changelog:
https://about.gitlab.com/releases/2021/10/28/security-release-gitlab-14-4-1-released/

Security:       33557582-3958-11ec-90ba-001b217b3468