lang/rust: Add riscv64 (riscv64gc-unknown-freebsd) support
A kernel with COMPAT_FREEBSD11 is required for now like on all
other FreeBSD archs. FreeBSD11 never had riscv64 support, so
GENERIC does not have it and a kernel rebuild is required. The
same issue existed on powerpc64le.

The riscv64gc-unknown-freebsd target was upstreamed and will
be part of Rust 1.59.0 (2021-02-24).

Reviewed by:	jrtc27, pkubaj
Differential Revision:	https://reviews.freebsd.org/D32881