games/gzdoom: update to 4.7.1
Changelogs for all of the intervening versions can be found on the
ZDoom forums: https://forum.zdoom.org/viewforum.php?f=1

This update moves data up to share/doom, which is a standard location
that gzdoom searches for game data already.  Some of the names are
fairly generic, but not likely to conflict with other doom ports.

While we're here, slim down the dependencies a little bit -- many of
them are not needed anymore.