www/libhsts: add port: Library to check a domain against the Chromium HSTS Preload list
The HSTS preload list is a list of domains that support HTTPS. The
list is compiled by Google and is utilised by Chrome, Firefox and
others.

With this information, a HTTP client may contact a website without
trying a plain-text HTTP connection first. It prevents interception
with redirects that take place over HTTP. None of the sent data
will ever be unencrypted.

WWW: https://gitlab.com/rockdaboot/libhsts