graphics/wings: update to latest git version and allow build
As the released version does not work with Erlang 24 and the
latest devel release is rather buggy, update to the latest commit
from github. That does not suggest that this version is perfect,
I observe some wonkiness when actually displaying models; but at
least it builds.

While here, simplify the port: it doesn't work any worse when
not copying and symlinking the installation directory.