www/gitlab-ce: fix segfaults
Accessing the project page or doing anything that is related to a license
file caused segfault in gitaly.
Upgrading now the gems to this specific version fixed the problem.
As we need to use a prerelease of rubygem-grpc, it is added as extra port.

Please see also upstream here:
https://gitlab.com/gitlab-org/gitlab/-/issues/345693

PR:		259848
Reported by:	admin@asu-vei.ru