security/gnupg: fix table / enum mismatch
PR259775 reports that (auto)configuration behaves weirdly.
This is caused by a mismatch between an enum in the code
and a table that expects to match the enum values. When
BUILD_WITH_TPM2D is off (the default; I have not looked if
it can even be turned on in ports) one table entry is

Patch also submitted upstream.

PR:		259775
Reported by:	O. Hartmann
MFH:		2021Q3