games/slade: pin wxWidgets version down to 3.0 (rather than 3.0+)
Version 3.1 defaults to EGL (presumably for Wayland) and is incompatible
with the GLEW we build at the moment (which defaults to GLX).  Basically,
because of this, the renderer craps out and crashes with v3.1.

We don't really know what the ramifications are of swapping all of it to
EGL, so just sticking with v3.0 seems safe for now.