games/r1q2: unbreak the build on recent -CURRENT
Something had been changed in our toolchain which broke linking of this
port because `SV_FixPlayerSkin' symbol became undefined.  This function
was annotated as `__inline', change it to `static' to unbreak the build
as it is defined and called within a single file only.

While here, remove unsupported optimization flag from the CFLAGS.

Reported by:	pkg-fallout