bsd.port.mk: Do not pass FLAVOR in dependency make environment
The change introduced in bug 256301 (review D30579), which prevents
child make processes from receiving an empty FLAVOR variable when
FLAVOR should be unset, has the side effect of allowing any FLAVOR
already in the parent make process environment to propagate to the
child.

This revision prevents the FLAVOR from the parent make from
incorrectly propagating to the child during a recursive make.

Additionally, use "flavored" CURDIR in recursive_dirs variable
so that a correct FLAVOR is used for the top-most port as well.

Reviewed by:	bapt
Approved by:	bapt
Differential Revision: https://reviews.freebsd.org/D32324