audio/praat: fix build on 14-
Build failures reported upstream at
	https://github.com/praat/praat/issues/1933
The build failure is in test-code, in a static function
that does nothing useful -- looks like some C++-experimentation
that is still in the source tree. Massage it away so that
builds on 14- can happen. Bump PORTREVISION since on pre-14-
it might now pick a different constructor, and so potentially
the package changes.