www/firefox: make dep on wasi-compiler-rt more flexible
as firefox needs matching wasi-compiler-rt, embedded llvm in
rust and llvm for building and wasi-compiler-rt installs it's files
in a path depending on it's full distversion, we need to know
llvm/clang's full version. it's not possible to derive that
programatically, as the required llvm version may not be installed
at the time firefox' Makefile is being parsed.

(cherry picked from commit dcc6b13daedf095da6572c27af69713ee551cb81)