editors/neovim: Update to 0.6.0 and switch lua backend
0.6.0 is an interval update that refines and extends the major lua
overhaul of 0.5.0.

Instead of using options for luajit vs PUC (mainline) lua, use
luajit-openresty unconditionally. devel/luv needs to use the same
backend as NeoVIM, so this commit switches it too.

NeoVIM is designed specifically to run on luajit rather than PUC;
upstream essentially doesn't develop for PUC. However, luajit is
largely unmaintained, locked in an older lua dialect, and unbuildable
on many architectures.

The luajit-openresty fork is faster, more modern, and should work in
more places. It's unclear to me which FreeBSD archs it does or doesn't
work on, so the fallout will have to be addressed once the builders have
a whack at it.