x11-fonts/libXft: Update to 2.3.4, latest upstream
This is mostly a build-cleanliness release, and prepares for some
fontconfig API changes. See release notes at
	https://gitlab.freedesktop.org/xorg/lib/libxft/-/tags/libXft-2.3.4
(well, go to the commit linked from there anyway).

NOT bumping PORTREVISION of consumers, since that would be a huge part
of the tree, and there are no bug fixes and no changes to behavior:
the calls to deprecated functions in fontconfig (the calls have been
removed in Xft) already did nothing on the fontconfig side.

PR:		260168
Reported by:	Dmitry Lukhtionov