emulators/vt100: Add port: Simulation of VT100 terminal hardware
This emulator emulates the internal circuitry and processor of the VT100 for a
very realistic user experience.

It is slightly patched to work smoothly on FreeBSD.  Some additional quality of
life patches have been added in discussion with upstream and will be added to a
future release.

This is a software simulation of the VT100 hardware. The original
firmware ROM is built in and executed by an 8080 emulator.  Other
components include video display with character generator ROM, settings
NVRAM, Intel 8251 USART, and a keyboard matrix scanner.  The Advance
Video Option is not included.

WWW: https://github.com/larsbrinkhoff/terminal-simulator

PR:		259956