devel/kf5-kconfig: add dependency on qt5-declarative
In 5.89 a component was moved from kf5-kdeclarative to kf5-kconfig,
but the activation is based on optionally installed qt5-declarative.

This might get moved to a default-on option in the future.

This also makes the optional qt5-declarative available transitively in
devel/kf5-knotifications, which therefore was also updated to
explicitely depend on that.

Reported by:	se, jkim