multimedia/pipewire: softer dependency on ncurses after 3af7fd92d368 (+)
Let framework to decide which ncurses to use on recent -RURRENT
if it still polluted with installed devel/ncurses as dependency
for other unadjusted ports.

This is artificial case but can be faced on machines which not rebuild
from scratch (or in a clean environment) after base system upgrades.

No PORTREVISION bump, clean env build (poudriere f.e.) is not affected.

Reported by:	rhurling
Sponsored by:	Netzkommune GmbH