x11-toolkits/qt5-quick3d: switch to use bundled assimp
Quick3d is at the moment not compatible with assimp-5.1.0. In poudriere,
due to assimp not being available, the port was already built with the
bundled assimp version.

For builds in a non-clean-room environment, explitely specify, that we
want the bundled one too.

PR:		260355
Reported by:	Bertrand Petit <bsdpr@phoe.frmug.org>