games/vvvvvv: update to 2.3.6
The game now comes in two flavours:

The commercial flavour is the regular game and requires the user to
supply his or her own copy of the data.zip file.

The makeandplay flavour is a free version of the game which doesn't
have come with the original levels, but ships player levels and
allows you to build your own.

PR:	260749