benchmarks/stress-ng: fix build on CURRENT,
In CURRENT, the initial configuration detects now support for the
scheduler affinity, enabling the build of an additional stressor.
In this stressor, the CPU_OR macro has 3 arguments (Linux implementation),
while on FreeBSD it has only 2 arguments.

This patch provides the fix

No bumprevision is provided, because it's only fixing CURRENT where the
build was failing.