graphics/darktable38: new v3.8 of raw developer
This port is meant to fill the gap until we can upgrade
graphics/darktable (which is v3.6), so that we have 3.8 in 2022Q1,
and to avoid breaking darktable for the latter in case new issues
show up, as darktable 3.8 is fairly new at the time of commit.

This is based on prior work by the graphics/darktable port maintainers,
dumbbell@ and lwhsu@.

Versus v3.6, we re-enable ninja builds but LTO with RawSpeed
currently does not work. For that to work, we need to BINARY_ALIAS
the linker, to avoid picking up the base system linker with a newer
LLVM toolchain from ports (no matter if 12 or 13).

related to
PR:		260674