sysutils/slurm-wlm: fix build on -CURRENT
The CPU_OR macro has been made compatible with GLIBC on -CURRENT,
the FreeBSD specific alternative is therefore no longer required.

While here pass --disable-slurmrestd to prevent the auto-configure
script from attempting to build slurmrestd if glib is found by
auto-configure (it does not build on FreeBSD due to Linux specific
headers and shell features provided by Bash but by the /bin/sh in
FreeBSD).

This port has numerous build issues when not building in a clean
jail (i.e. on a system with a previous version installed or with
other software that is detected by auto-configure).

The port revision is not bumped since these changes do not modify
the package that is generated.