audio/codec2: Unbreak swig
Numerous tiny changes in the git log for codec2 but
the big one for us is fixing swig.
codec2 exposed a complex type in an interface that swig breaks on.
This impacts other ports such as gnuradio.
upstream moved the interface to private and this unbreaks swig.
N.B. I have patched out the Octave CML testjig for now
since this will take more time to resolve.
Patches welcome.