converters/iconv-rfc1345: add license
There is no indication of a license intended by the author, but
since this is an extension of the converters/iconv port which is
distributed under a 3-clause BSD license, add it to this port, too.

While here use @preunexec instead of the deprecated @unexec in
pkg-plist.

Approved by:	bland (maintainer)