emulators/qemu-user-static-devel: Update to latest devel version
Update to the latest pst 6.2 qemu development version for bsd-user.
This removes mips support and includes new system calls like swapoff
with two args. It's also the result of continuing efforts to upstream
changes. It's a devel version, and more dicy than most devel versions,
so caution is advised: it builds but there may be important cases that
are broken.

Sponsored by:		Netflix