devel/rubygem-readline: Add rubygem-readline 0.0.3
This is just a loader for "readline". If Ruby has the "readline-ext" gem that is
a native extension, this gem will load it. If Ruby does not have the
"readline-ext" gem this gem will load "reline", a library that is compatible
with the "readline-ext" gem and implemented in pure Ruby.

WWW: https://github.com/ruby/readline